+++ CardinalKit-Example/CardinalKit/Components/MainUIView.swift
@@ -1,0 +1,41 @@
+//
+//  MainUIView.swift
+//  CardinalKit_Example
+//
+//  Created for the CardinalKit Framework.
+//  Copyright © 2019 Stanford University. All rights reserved.
+//
+
+import SwiftUI
+
+struct MainUIView: View {
+    
+    let color: Color
+    let config = CKPropertyReader(file: "CKConfiguration")
+    
+    init() {
+        self.color = Color(config.readColor(query: "Primary Color"))
+    }
+    
+    var body: some View {
+        TabView {
+            StudyActivitiesUIView(color: self.color)
+                .tabItem {
+                    Image("tab_activities").renderingMode(.template)
+                    Text("Activities")
+            }
+
+            ProfileUIView(color: self.color)
+                .tabItem {
+                    Image("tab_profile").renderingMode(.template)
+                    Text("Profile")
+                }
+        }.accentColor(self.color)
+    }
+}
+
+struct MainUIView_Previews: PreviewProvider {
+    static var previews: some View {
+        MainUIView()
+    }
+}
