+++ CardinalKit-Example/CardinalKit/Components/Profile/Views/ChangePasscodeView.swift
@@ -1,0 +1,31 @@
+//
+//  ChangePasscodeView.swift
+//  CardinalKit_Example
+//
+//  Created for the CardinalKit Framework.
+//  Copyright © 2019 Stanford University. All rights reserved.
+//
+
+import SwiftUI
+import ResearchKit
+
+struct ChangePasscodeView: View {
+    @State var showPasscode = false
+    
+    var body: some View {
+        HStack {
+            Text("Change Passcode")
+            Spacer()
+            Text("›")
+        }.frame(height: 70).contentShape(Rectangle())
+            .gesture(TapGesture().onEnded({
+                if ORKPasscodeViewController.isPasscodeStoredInKeychain() {
+                    self.showPasscode.toggle()
+                }
+        })).sheet(isPresented: $showPasscode, onDismiss: {
+            
+        }, content: {
+            PasscodeViewController()
+        })
+    }
+}
