+++ CardinalKit-Example/CardinalKit/Components/Profile/Views/HelpView.swift
@@ -1,0 +1,30 @@
+//
+//  HelpView.swift
+//  CardinalKit_Example
+//
+//  Created for the CardinalKit Framework.
+//  Copyright © 2019 Stanford University. All rights reserved.
+//
+
+import SwiftUI
+
+struct HelpView: View {
+    var site = ""
+    
+    init(site: String) {
+        self.site = site
+    }
+    
+    var body: some View {
+        HStack {
+            Text("Help")
+            Spacer()
+            Text("›")
+        }.frame(height: 70).contentShape(Rectangle())
+            .gesture(TapGesture().onEnded({
+                if let url = URL(string: self.site) {
+                UIApplication.shared.open(url)
+            }
+        }))
+    }
+}
