+++ CardinalKit-Example/CardinalKit/DEPRECATED/_ProfileTableItem.swift
