+++ CardinalKit-Example/CardinalKit/DEPRECATED/_StaticProfileViewController.swift
