+++ CardinalKit-Example/CardinalKit/DEPRECATED/_StudyTableViewCell.swift
