+++ CardinalKit-Example/CardinalKit/Utils/ConsentDocument.swift
