+++ CardinalKit-Example/CardinalKit/Utils/EmailHelper.swift
@@ -1,0 +1,36 @@
+//
+//  EmailHelper.swift
+//  CardinalKit_Example
+//
+//  Created by Santiago Gutierrez on 10/12/20.
+//  Copyright © 2020 CocoaPods. All rights reserved.
+//
+
+import MessageUI
+
+class EmailHelper: NSObject, MFMailComposeViewControllerDelegate {
+    public static let shared = EmailHelper()
+
+    func sendEmail(subject:String, body:String, to:String){
+        if !MFMailComposeViewController.canSendMail() {
+            return
+        }
+        
+        let picker = MFMailComposeViewController()
+        
+        picker.setSubject(subject)
+        picker.setMessageBody(body, isHTML: true)
+        picker.setToRecipients([to])
+        picker.mailComposeDelegate = self
+        
+        EmailHelper.getRootViewController()?.present(picker, animated: true, completion: nil)
+    }
+    
+    func mailComposeController(_ controller: MFMailComposeViewController, didFinishWith result: MFMailComposeResult, error: Error?) {
+        EmailHelper.getRootViewController()?.dismiss(animated: true, completion: nil)
+    }
+    
+    static func getRootViewController() -> UIViewController? {
+        (UIApplication.shared.connectedScenes.first?.delegate as? SceneDelegate)?.window?.rootViewController
+    }
+}
